+++ tests/e2e/api/src/models/products/abstract/index.ts
@@ -1,6 +1,7 @@
 export * from './common';
 export * from './cross-sell';
 export * from './delivery';
+export * from './external';
 export * from './grouped';
 export * from './inventory';
 export * from './price';