+++ tests/e2e/api/src/models/products/index.ts
@@ -2,3 +2,4 @@ export * from './abstract';
 export * from './shared';
 export * from './simple-product';
 export * from './grouped-product';
+export * from './external-product';