+++ tests/e2e/api/src/repositories/rest/products/shared.ts
@@ -14,6 +14,7 @@ import {
 	AbstractProduct,
 	IProductCrossSells,
 	IProductDelivery,
+	IProductExternal,
 	IProductGrouped,
 	IProductInventory,
 	IProductSalesTax,
@@ -189,6 +190,9 @@ export function createProductTransformer< T extends AbstractProduct >(
 	return new ModelTransformer( transformations );
 }
 
+/**
+ * Create a transformer for the product cross sells property.
+ */
 export function createProductCrossSellsTransformation(): ModelTransformation[] {
 	const transformations = [
 		new PropertyTypeTransformation(
@@ -206,6 +210,9 @@ export function createProductCrossSellsTransformation(): ModelTransformation[] {
 	return transformations;
 }
 
+/**
+ * Create a transformer for the product upsells property.
+ */
 export function createProductUpSellsTransformation(): ModelTransformation[] {
 	const transformations = [
 		new PropertyTypeTransformation(
@@ -243,6 +250,9 @@ export function createProductGroupedTransformation(): ModelTransformation[] {
 	return transformations;
 }
 
+/**
+ * Create a transformer for product delivery properties.
+ */
 export function createProductDeliveryTransformation(): ModelTransformation[] {
 	const transformations = [
 		new ModelTransformerTransformation( 'downloads', ProductDownload, createProductDownloadTransformer() ),
@@ -269,6 +279,9 @@ export function createProductDeliveryTransformation(): ModelTransformation[] {
 	return transformations;
 }
 
+/**
+ * Create a transformer for product inventory properties.
+ */
 export function createProductInventoryTransformation(): ModelTransformation[] {
 	const transformations = [
 		new PropertyTypeTransformation(
@@ -298,6 +311,9 @@ export function createProductInventoryTransformation(): ModelTransformation[] {
 	return transformations;
 }
 
+/**
+ * Create a transformer for product sales tax properties.
+ */
 export function createProductSalesTaxTransformation(): ModelTransformation[] {
 	const transformations = [
 		new PropertyTypeTransformation(
@@ -317,6 +333,9 @@ export function createProductSalesTaxTransformation(): ModelTransformation[] {
 	return transformations;
 }
 
+/**
+ * Create a transformer for product shipping properties.
+ */
 export function createProductShippingTransformation(): ModelTransformation[] {
 	const transformations = [
 		new CustomTransformation(
@@ -369,3 +388,26 @@ export function createProductShippingTransformation(): ModelTransformation[] {
 
 	return transformations;
 }
+
+/**
+ * Transformer for the properties unique to the external product type.
+ */
+export function createProductExternalTransformation(): ModelTransformation[] {
+	const transformations = [
+		new PropertyTypeTransformation(
+			{
+				buttonText: PropertyType.String,
+				externalUrl: PropertyType.String,
+			},
+		),
+		new KeyChangeTransformation< IProductExternal >(
+			{
+				buttonText: 'button_text',
+				externalUrl: 'external_url',
+			},
+		),
+	];
+
+	return transformations;
+}
+