+++ tests/e2e/core-tests/specs/index.js
@@ -8,7 +8,8 @@ const runActivationTest = require( './activate-and-setup/activate.test' );
 const { runOnboardingFlowTest, runTaskListTest } = require( './activate-and-setup/onboarding-tasklist.test' );
 const runInitialStoreSettingsTest = require( './activate-and-setup/setup.test' );
 
-// API tests
+// REST API tests
+const runExternalProductAPITest = require( './api/external-product' );
 const runGroupedProductAPITest = require( './api/grouped-product' );
 
 // Shopper tests
@@ -18,10 +19,12 @@ const runCheckoutApplyCouponsTest = require( './shopper/front-end-checkout-coupo
 const runCheckoutPageTest = require( './shopper/front-end-checkout.test' );
 const runMyAccountPageTest = require( './shopper/front-end-my-account.test' );
 const runSingleProductPageTest = require( './shopper/front-end-single-product.test' );
+const runVariableProductUpdateTest = require( './shopper/front-end-variable-product-updates.test' );
 
 // Merchant tests
 const runCreateCouponTest = require( './merchant/wp-admin-coupon-new.test' );
 const runCreateOrderTest = require( './merchant/wp-admin-order-new.test' );
+const runEditOrderTest = require( './merchant/wp-admin-order-edit.test' );
 const { runAddSimpleProductTest, runAddVariableProductTest } = require( './merchant/wp-admin-product-new.test' );
 const runUpdateGeneralSettingsTest = require( './merchant/wp-admin-settings-general.test' );
 const runProductSettingsTest = require( './merchant/wp-admin-settings-product.test' );
@@ -47,11 +50,13 @@ const runShopperTests = () => {
 	runCheckoutPageTest();
 	runMyAccountPageTest();
 	runSingleProductPageTest();
+	runVariableProductUpdateTest();
 };
 
 const runMerchantTests = () => {
 	runCreateCouponTest();
 	runCreateOrderTest();
+	runEditOrderTest();
 	runAddSimpleProductTest();
 	runAddVariableProductTest();
 	runUpdateGeneralSettingsTest();
@@ -71,6 +76,7 @@ module.exports = {
 	runTaskListTest,
 	runInitialStoreSettingsTest,
 	runSetupOnboardingTests,
+	runExternalProductAPITest,
 	runGroupedProductAPITest,
 	runCartApplyCouponsTest,
 	runCartPageTest,
@@ -78,9 +84,11 @@ module.exports = {
 	runCheckoutPageTest,
 	runMyAccountPageTest,
 	runSingleProductPageTest,
+	runVariableProductUpdateTest,
 	runShopperTests,
 	runCreateCouponTest,
 	runCreateOrderTest,
+	runEditOrderTest,
 	runAddSimpleProductTest,
 	runAddVariableProductTest,
 	runUpdateGeneralSettingsTest,