+++ tests/e2e/specs/front-end/test-variable-product-updates.js
@@ -1,0 +1,6 @@
+/*
+ * Internal dependencies
+ */
+const { runVariableProductUpdateTest } = require( '@woocommerce/e2e-core-tests' );
+
+runVariableProductUpdateTest();