+++ tests/e2e/api/src/models/products/index.ts
@@ -1,4 +1,5 @@
 export * from './abstract';
 export * from './shared';
 export * from './simple-product';
+export * from './grouped-product';
 export * from './external-product';