+++ tests/e2e/api/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Added
 
 - Support for the external product type.
+- Support for coupons.
 
 ## Breaking Changes
 
@@ -18,7 +19,7 @@
 
 ## Changes
 
-- Added a tranformation layer between API responses and internal models
+- Added a transformation layer between API responses and internal models
 
 ## Fixed