+++ tests/e2e/api/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Support for the external product type.
 - Added support for grouped product type
+- Support for coupons.
 
 # 0.1.1
 
@@ -19,7 +20,7 @@
 
 ## Changes
 
-- Added a tranformation layer between API responses and internal models
+- Added a transformation layer between API responses and internal models
 
 ## Fixed